Update spec const tests: move user-def structs into common folder
